fix grade exam

+++ course-service/course-domain/course-application-service/src/main/java/com/backend/programming/learning/system/course/service/domain/dto/method/query/exam/QueryGradeResponse.java
@@ -21,7 +21,7 @@ public record QueryGradeResponse(
         String status,
         ZonedDateTime lastSubmitAt,
         ZonedDateTime lastMarkAt,
-        Float score,
-        Float maxScore
+        Double score,
+        Double maxScore
 ) {
 }
